+++ packages/expo-camera/ios/Current/CameraPhotoCapture.swift
@@ -188,7 +188,6 @@ class CameraPhotoCapture: NSObject, AVCapturePhotoCaptureDelegate {
     let croppedSize = AVMakeRect(aspectRatio: previewSize, insideRect: cropRect)
 
     takenImage = ExpoCameraUtils.crop(image: takenImage, to: croppedSize)
-    takenImage = ExpoCameraUtils.normalizeOrientation(of: takenImage)
 
     let width = takenImage.size.width
     let height = takenImage.size.height
@@ -206,12 +205,15 @@ class CameraPhotoCapture: NSObject, AVCapturePhotoCaptureDelegate {
         with: ["Orientation": ExpoCameraUtils.toExifOrientation(orientation: takenImage.imageOrientation)]
       )
 
-      updatedExif[kCGImagePropertyExifPixelYDimension as String] = width
-      updatedExif[kCGImagePropertyExifPixelXDimension as String] = height
+      if let cgImage = takenImage.cgImage {
+        updatedExif[kCGImagePropertyExifPixelXDimension as String] = cgImage.width
+        updatedExif[kCGImagePropertyExifPixelYDimension as String] = cgImage.height
+      }
       response["exif"] = updatedExif
 
       var updatedMetadata = metadata
-      updatedMetadata[kCGImagePropertyOrientation as String] = 1
+      updatedMetadata[kCGImagePropertyOrientation as String] =
+        ExpoCameraUtils.toExifOrientation(orientation: takenImage.imageOrientation)
 
       if let additionalExif = options.additionalExif {
         for (key, value) in additionalExif {